+++ internal/commands/tmmeta.go
@@ -11,7 +11,7 @@ import (
 func GetTMMetadata(ctx context.Context, spec model.RepoSpec, tmID string) (*model.FoundVersion, error) {
 	repo, err := repos.Get(spec)
 	if err != nil {
-		return nil, fmt.Errorf("Could not ìnitialize a repo instance for %s: %w\ncheck config", spec, err)
+		return nil, fmt.Errorf("could not ìnitialize a repo instance for name %s: %w", spec, err)
 	}
 
 	sr, err := repo.GetTMMetadata(ctx, tmID)